adaptlearn/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-dotenv
google-generativeai
python-jose[cryptography]
passlib[bcrypt]
google-auth
python-multipart
httpx
requests
